pyproject.toml
[project]
name = "digital-hub"
version = "0.1.0"
description = ""
authors = [
    {name = "Mikiyas Degefu",email = "mikiyas.m.degefu@gmail.com"},
    {name = "Kaleab Hegie",email = "benjiyg400@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "django (==4.2.6)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)",
    "pymupdf (>=1.25.3,<2.0.0)",
    "rest-framework-simplejwt (>=0.0.2,<0.0.3)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "django-ckeditor (>=6.7.2,<7.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "firebase-admin (>=7.1.0,<8.0.0)",
    "django-axes[ipware] (>=8.0.0,<9.0.0)",
    "python-socketio[asgi] (>=5.15.0,<6.0.0)",
    "celery (>=5.6.0,<6.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "uwsgi (>=2.0.31,<3.0.0)",
    "langchain (>=1.1.3,<2.0.0)",
    "langchain-milvus (>=0.3.1,<0.4.0)",
    "langchain-text-splitters (>=1.0.0,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "langchain-classic (>=1.0.0,<2.0.0)",
    "vllm (>=0.12.0,<0.13.0)",
    "langchain-openai (>=1.1.3,<2.0.0)",
    "sentence-transformers (>=5.2.0,<6.0.0)",
    "langchain-huggingface (>=1.2.0,<2.0.0)",
    "langchain-core (>=1.2.0,<2.0.0)",
    "channels[daphne] (>=4.3.2,<5.0.0)",
    "daphne (>=4.2.1,<5.0.0)",
    "channels-redis (>=4.3.0,<5.0.0)",
    "drf-user-activity-tracker (>=1.0.1,<2.0.0)",
    "aiofiles (>=25.1.0,<26.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
